Fix Orca CTAS for tables created with legacy hashops
For CTAS creating tables with non-legacy hashop distribution from tables
with legacy hashops, CTAS with Orca would distribute the data according
to the value of gp_use_legacy_hashops; however, it would set the table's
distribution policy hashop to the value of the original table. This
caused queries to give incorrect results as the distribution policy
mismatched the data distribution.

Now, we also set the new table's distribution policy according to the
opclass passed into the query and roundtrip this through Orca.
This is aligned with the planner behavior and the table's actual
distribution.