Avoid ".." in include paths to accommodate tooling.
The current setup in CMake (and Makefile's too, but that's an even
harder problem) leads to equivalent-but-not-identical header include
paths (-I) for the same directory, e.g. -Ilibgpopt/include vs
-Ilibgpdbcost/../libgpopt/include .

This confuses Clang-based tooling into identifying multiple paths for
the same header -- the difference being the extra sibling directory
followed by dot-dot, or "niece" directory followed by two levels of
dot-dot, and so forth -- as different headers. That, in turn, undermines
the conflict resolution and edit deduplication features in Clang's
refactoring engine, leading to duplicate edits when applying FixIt's.

This commit applies some fairly simple fixes to spell the sibling
directories in a way that generates consistent include paths, so that
Clang tooling are more functional. The Makefile's are left unchanged as
that is a lot more difficult to make "right". One can argue that we
_might_ want to instead transform the intermediate representation of
Clang's "replacement" YAML files, but that's left for another day.