Publicize deleted member functions
Now that we are explicitly declaring copy-assignment operators and copy
constructors as deleted, we should also make them public -- private and
=delete don't make much sense in combination, and this results in the
best diagnostics in practice [1].

In the previous commit where private unimplemented special member
functions are changed to be declared as deleted, we left the new
declarations private. This follows up by moving those deleted functions
to the public section of their classes.

I made this change guided by tooling: even though it doesn't offer a
fix, clang-tidy is useful enough to emit diagnostics that look like the
following:

/home/pivotal/workspace/gpdb/src/backend/gporca/libgpopt/include/gpopt/base/CColRef.h:76:2: warning: deleted member function should be public [modernize-use-equals-delete]
        CColRef(const CColRef &) = delete;
        ^

The process involved of making this change is a tale of much sorrow,
blood and tears. Suffice it to say 20 different regular expressions were
involved, and I'm edging closer to a PTSD when looking at backslashes.

References:
[1] https://abseil.io/tips/143#summary