Add Casting.h
This provides a uniform interface for runtime casting pointers and
Ref<T> objects. The implementation is a little rudimentary (it's just
dynamic_cast right now), but it already allows "moving" casts. This
should allow us a more performant implementation of casting should we
decide to investing in that down the road (hint: no dynamic_cast).