Manually mark one pointer
This method is temporarily dead because of ongoing partitioning work.
Meanwhile, mark it as a pointer manually. After removing two uber
headers "ops.h" and "dxlops.h", we no longer see the AddRef() and
Release() methods instantiated for the type "IStatisticsArray", so our
refactoring tool cannot recognize it as a reference counted type.