Consistently use "owner" fields in constructors
While converting this code base towards using a smart pointer-based
approach to manage object lifetime, we noticed that a big optimization
opportunity missed in the converted code is the following pattern: a
class C "owns" a field m_f, and in a constructor for C we initialize m_f
with a parameter f, assuming the caller has transferred the ownership:

C::C(T *f): m_f(f) {...}

The desired converted code should also transfer ownership by applying
std::move to the parameter:

C::C(Ref<T> f): m_f(std::move(f)) {...}

but it doesn't. The root cause is the parameter is still used in the
body of the constructor, often for an assertion, so the conversion
program gives up out of caution (otherwise we run into use-after-move).

This commit finds such pattern and changes the code to equivalently use
the initialized field instead of the parameter.